requirements.txt
Flask>=2.2.5
python-dotenv>=1.0.0
openai>=1.3.0
Jinja2>=3.1.2
PyYAML>=6.0
jsonschema>=4.20.0
azure-identity>=1.14.0
azure-mgmt-datafactory>=1.1.0
Flask-Cors>=4.0.0
azure-mgmt-subscription>=3.1.1
azure-mgmt-resource>=23.1.1
azure-mgmt-storage>=21.1.0
azure-storage-blob>=12.22.0